Begin Capturing State, test client
WARNING! This leaks memory until we decide how we want to deal with some
kinds of deletion. The MallocMessageBuilder pointers I create called
neverFree and then allow to go out of scope are required to release the
Facts or FactSigs that are living in the database. If MemDAL ever ends
up being anything other than a test component, this will need to be
fixed.