Make macros more permissive about use of iter
Previously, I was trying to make the macros avoid the situation where a
destructure around an iteration would happen, since while it would work,
the semantics could easily be surprising.
However, the way I did it caused it to be impossible to destructure and
iterate in the same where clause, which is incredibly important, so I've
lifted the restriction temporarily.

Overall, this is suggesting that I may want to move my macro language
into a compiler plugin for more flexibilit yin the future.