Fix Nix tooling for library mode
The Rust build support in nix is oriented around building binaries, not
libraries. As a result, it requires the presence of Cargo.lock, which
explicitly should not be a part of Rust libraries. I've replaced
default.nix/package.nix with a shell.nix for development, since rust
libraries are not separately installable.