Change to hybrid of btree and hash indexing
I switched to hash-uniquing to allow extending values (e.g. traces) to
be uniqued without local code support, but this caused a major
performance drop due to the lack of indices on other fields during
joins.

As a middle ground, I'm now adding a unique index on a tuple of all
columns, but columns marked "large" are hashed before being placed in
the btree to prevent them from growing too large for an index page.